pyproject.toml
[tool.poetry]
name = "news-shepherd"
version = "0.1.0"
description = "Backed that will parse user chosen telegram channels and apply RAG to 'em to find most relatable items from those, maybe event filter them out"
authors = ["parabolabam <sorokinvladislav1402@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
transformers = "^4.46.1"
torch = "^2.5.1"
python-telegram-bot = "^21.6"
supabase = "^2.9.1"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
telegram = "^0.0.1"
python-dotenv = "^1.0.1"
telethon = "^1.38.1"
openai = "^1.55.0"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
